refactor: apply architecture
