+++ src/auth/models/domain/tokens.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class TokenData(BaseModel):
+    """ 토큰 데이터 """
+
+    username: str
