+++ src/auth/repository/fake.py
@@ -1,0 +1,16 @@
+users = {
+    "mcauto": {
+        "username": "mcauto",
+        "full_name": "Mincheol Kim",
+        "email": "nexters@kakao.com",
+        "hashed_password": "fakehashedsecret",
+        "disabled": False,
+    },
+    "guest": {
+        "username": "guest",
+        "full_name": "Player Unknown",
+        "email": "guest@guest.guest",
+        "hashed_password": "fakehashedsecret",
+        "disabled": True,
+    },
+}
