+++ src/routes/__init__.py
@@ -4,8 +4,8 @@ routes
 from fastapi.routing import APIRouter
 
 from .health import health as health_router
-from .todos import todos as todos_router
-from .token import token as token_router
+from ..todos.routes import todos as todos_router
+from ..auth.routes import token as token_router
 
 __all__ = ["api_v1", "health_router", "todos_router", "token_router"]
