+++ src/todos/models/domain/todos.py
@@ -1,0 +1,11 @@
+from typing import Optional
+from pydantic import BaseModel
+
+
+class TodoItem(BaseModel):
+    """ todo schema """
+
+    seqno: int
+    title: str
+    content: Optional[str]
+    is_done: bool
