+++ src/todos/repository/fake.py
@@ -1,0 +1,4 @@
+from typing import Dict
+from ..models.domain.todos import TodoItem
+
+todos: Dict[int, TodoItem] = {}
